Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0.post1
alembic==1.13.0
pytest==7.4.3
httpx==0.25.2
slowapi==0.1.8
cachetools==5.3.2
aiosqlite==0.19.0
asyncpg==0.29.0
greenlet==3.0.2
